requirements.txt
absl-py==1.1.0
antlr4-python3-runtime==4.9.3
chex==0.1.3
cycler==0.11.0
dm-haiku==0.0.7
dm-tree==0.1.7
etils==0.6.0
flatbuffers==2.0
fonttools==4.34.4
hydra-core==1.2.0
importlib-resources==5.8.0
jax==0.3.14
jaxlib==0.3.14
jmp==0.0.2
kiwisolver==1.4.3
matplotlib==3.5.2
numpy==1.23.1
omegaconf==2.2.2
opt-einsum==3.3.0
optax==0.1.3
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
scipy==1.8.1
six==1.16.0
tabulate==0.8.10
toolz==0.12.0
typing_extensions==4.3.0
zipp==3.8.1
